Fix Date/Timezone handling
The CAKE API returns timestamps with the local timezone offset but
without specifying the zone. Instead of having a fixed offset parameter,
parse and generate the CAKE dates with a user-specified timezone.

The old and broken way of parsing dates is still present and is active
as long as CAKE_TIME_OFFSET is set, so this should not break any
existing code. All code should be migrated to use the new CAKE_TIME_ZONE
parameter (and thus enabling the new, correct way of parsing dates)
as soon as possible.